Split types into files
